Introduced GetIt and extracted GameRepository
