+++ lib/getIt.dart
@@ -1,0 +1,9 @@
+import 'package:get_it/get_it.dart';
+
+import 'game/game_repository.dart';
+
+final getIt = GetIt.instance;
+
+void initialiazeGetItContext(){
+  getIt.registerSingleton(GameRepository());
+}
