+++ lib/main.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
+import 'getIt.dart';
 import 'spielwiesn_app.dart';
 
 void main() {
+  initialiazeGetItContext();
   runApp(const SpielwiesnApp());
 }
