+++ pubspec.yaml
@@ -19,6 +19,7 @@ dependencies:
   upgrader: ^12.1.0
   markdown_widget: ^2.3.2+8
   url_launcher: ^6.3.2
+  get_it: ^9.2.1
 
 dev_dependencies:
   flutter_test:
